requirements.txt
python-telegram-bot==20.7
mistralai==0.0.10
python-dotenv==1.0.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.103.0
firebase-admin==6.2.0
pillow==10.0.0